[HER-1843] Possible int overflow in CheckpointService.java * CheckpointService   Made checkpointIntervalMinutes setter accept a long instead of an int (was missed when the class variable was made a long). While it was non-harmful the way it was, this is more in line with Java conventions.

+++ engine/src/main/java/org/archive/crawler/framework/CheckpointService.java
@@ -92,7 +92,7 @@ public class CheckpointService implements Lifecycle, ApplicationContextAware, Ha
     public long getCheckpointIntervalMinutes() {
         return checkpointIntervalMinutes;
     }
-    public void setCheckpointIntervalMinutes(int interval) {
+    public void setCheckpointIntervalMinutes(long interval) {
         long oldVal = checkpointIntervalMinutes; 
         this.checkpointIntervalMinutes = interval;
         if(checkpointIntervalMinutes!=oldVal) {
